rsg_construction/requirements.txt
absl-py==1.4.0
annoy==1.17.1
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorcet==3.0.1
contourpy==1.0.7
cycler==0.11.0
dtaidistance==2.3.10
dtw==1.4.0
fonttools==4.39.2
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
neo4j==5.6.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.0
pacmap==0.7.0
pandas==1.5.3
param==1.13.0
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyct==0.5.0
pynndescent==0.5.8
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensordict==0.1.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torch-tb-profiler==0.4.1
tqdm==4.65.0
trimap==1.1.4
typing_extensions==4.5.0
umap-learn==0.5.3
urllib3==1.26.15
Werkzeug==2.2.3
yapf==0.32.0
zipp==3.15.0
pyyaml==5.3.1